Added in tests for recipie instruction

+++ src/main/java/com/amgenz/entity/RecipeIngredient.java
@@ -24,7 +24,7 @@ public class RecipeIngredient {
     private String ingredient;
 
     @Column(name = "ingredient_amount")
-    private int ingredientAmount;
+    private double ingredientAmount;
 
     @Column(name = "ingredient_amount_measurement")
     private String ingredientAmountMeasurement;
@@ -43,7 +43,7 @@ public class RecipeIngredient {
      * @param ingredientAmount            the ingredient amount
      * @param ingredientAmountMeasurement the ingredient amount measurement
      */
-    public RecipeIngredient(Recipe recipe, String ingredient, int ingredientAmount, String ingredientAmountMeasurement) {
+    public RecipeIngredient(Recipe recipe, String ingredient, double ingredientAmount, String ingredientAmountMeasurement) {
         this.recipe = recipe;
         this.ingredient = ingredient;
         this.ingredientAmount = ingredientAmount;
@@ -109,7 +109,7 @@ public class RecipeIngredient {
      *
      * @return the ingredient amount
      */
-    public int getIngredientAmount() {
+    public double getIngredientAmount() {
         return ingredientAmount;
     }
 
@@ -118,7 +118,7 @@ public class RecipeIngredient {
      *
      * @param ingredientAmount the ingredient amount
      */
-    public void setIngredientAmount(int ingredientAmount) {
+    public void setIngredientAmount(double ingredientAmount) {
         this.ingredientAmount = ingredientAmount;
     }
